Model and DAO impl

+++ ptpb-db/src/main/java/net/thumbtack/ptpb/db/project/ProjectMapper.java
@@ -2,6 +2,6 @@ package net.thumbtack.ptpb.db.project;
 
 import org.springframework.data.aerospike.repository.AerospikeRepository;
 
-interface ProjectMapper extends AerospikeRepository<Project, Integer> {
+interface ProjectMapper extends AerospikeRepository<Project, Long> {
 
 }
